Get data command

+++ database/migrations/2021_01_21_065348_create_payments_table.php
@@ -16,11 +16,11 @@ class CreatePaymentsTable extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->string('meter_id', 15);
-            $table->float('last_month', 5,2);
-            $table->float('this_month', 5,2);
-            $table->float('usage', 5,2);
-            $table->decimal('tariff', 10,2);
-            $table->decimal('penalty', 10,2)->nullable();
+            $table->float('last_month');
+            $table->float('this_month');
+            $table->float('usage');
+            $table->double('tariff');
+            $table->float('penalty')->nullable();
             $table->date('billing_month');
             $table->boolean('status')->default(1);
             $table->timestamps();
